// src/types/graphql-generated/graphql.d.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type AuthPayload = {
  __typename?: 'AuthPayload';
  token: Maybe<Scalars['String']>;
  userErrors: Array<UserError>;
};

export type ConfirmPayload = {
  __typename?: 'ConfirmPayload';
  confirmed: Maybe<Scalars['Boolean']>;
  userErrors: Array<UserError>;
};
// ...
export type MutationConfirmUserArgs = {
  token: Scalars['String'];
};
// ...
export type UserError = {
  __typename?: 'UserError';
  message: Scalars['String'];
  values: Maybe<Array<Maybe<Scalars['String']>>>;
};